Review fix: avoid signed/unsigned comparision in the assert

+++ dali/pipeline/executor/executor2/exec_node_task.cc
@@ -576,8 +576,8 @@ tasking::SharedTask ExecNodeTask::CreateTask(ExecNode *node, const WorkspacePara
   } else {
     int nout = node->outputs.size();
     // the size of outputs is limited by the compile-time operator schema
-    // that uses int for to represent the number of outputs
-    assert(nout >= 0 && nout == node->outputs.size());
+    // that uses int to represent the number of outputs
+    assert(nout >= 0 && static_cast<size_t>(nout) == node->outputs.size());
     return tasking::Task::Create(
       nout,
       OpTask(node, params).GetRunnable());
